tests/integeration_db/requirements.txt
pytest
docker
python-dotenv
SQLAlchemy==2.0.12